+++ app/services/forest_liana/base_getter.rb
@@ -27,9 +27,7 @@ module ForestLiana
     private
 
     def compute_includes
-      @includes = SchemaUtils.one_associations(@resource)
-        .select { |association| SchemaUtils.model_included?(association.klass) }
-        .map(&:name)
+      @includes = ForestLiana::QueryHelper.get_one_association_names_symbol(@resource)
     end
   end
 end