+++ spec/dummy/app/models/user.rb
@@ -1,0 +1,4 @@
+class User < ActiveRecord::Base
+  has_many :trees_owned, class_name: 'Tree', inverse_of: :owner
+  has_many :trees_cut, class_name: 'Tree', inverse_of: :cutter
+end